модули game и stats наследуют методы класса AbstractView
